[F] Align Linux distro detection with upstream Neofetch's behavior
The previous behavior of sourcing /etc/os-release *before*
/usr/lib/os-release causes janks especially when we need to spoof
our machine as another distro by altering /etc/os-release, such as
Ubuntu LTS for Microsoft Intune enrollment. Most of the apps that
rely on os-release's presence always rely on the one in /etc/,
skipping the one in /usr/lib/, so it's safe to assume this will
make the distro detection more reliable.

The behavior prior to this change is given in the asciinema recording
below:
Link: https://asciinema.org/a/647313

Signed-off-by: Stella Bloom <windowz414@gnuweeb.org>